+++ MoeaBench/plotting/__init__.py
@@ -3,7 +3,7 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .plotter import spaceplot, timeplot
+
 from .scatter3d import Scatter3D
 from .scatter2d import Scatter2D
 from .coordinate import ParallelCoordinate